[Jetsurvey] extract questions into their own kotlin files for readibility
